InventoryPRT/inventario_backend/requirements.txt
# inventario_backend/requirements.txt
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.29.0,<0.31.0
sqlalchemy>=2.0.0,<2.1.0
psycopg2-binary>=2.9.0,<2.10.0  # Si usas PostgreSQL
# mysqlclient>=2.2.0,<2.3.0    # Si usas MySQL (descomenta si aplica)
python-dotenv>=1.0.0,<1.1.0
passlib[bcrypt]>=1.7.0,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
python-multipart>=0.0.9,<0.0.10 # Para OAuth2PasswordRequestForm
pydantic[email]>=2.0.0,<2.9.0